SA-0MLJW08I01VIZ6JB: Rename ampa/metrics.py to ampa/server.py
The module serves /metrics, /health, /respond, /admin/fallback, and /session
endpoints — 'metrics' was misleading. Updated all imports, docstring, logger
name, and README references. All 765 tests pass.
